profiale serializer structure
